Views and Controllers updated to new patterns
