+++ ckanext/security/views.py
@@ -1,36 +1,46 @@
 # -*- coding: utf-8 -*-
 
-import ckantoolkit as tk
+from ckan.views import user
+from ckanext.security import utils
 from flask import Blueprint, make_response
+from functools import wraps
 
-security = Blueprint("security", __name__)
+
+def login_required(f):
+    @wraps(f)
+    def decorated_function(*args, **kwargs):
+        utils.check_user_and_access()
+        return f(*args, **kwargs)
+    return decorated_function
 
 
-@security.before_request
-def before_request():
+mfa_user = Blueprint("mfa_user", __name__)
 
-    # def before_map(self, urlmap):
-    #     userController = 'ckanext.security.controllers:SecureUserController'
-    #     urlmap.redirect('/user/edit/', '/user/edit')
-    #     urlmap.connect('/user/edit', controller=userController,
-    #                    action='edit')
-    #     urlmap.connect('/user/edit/{id:.*}', controller=userController,
-    #                    action='edit', ckan_icon='cog')
-    #     urlmap.connect('/user/reset/{id:.*}', controller=userController,
-    #                    action='perform_reset')
-    #     urlmap.connect('/user/reset', controller=userController,
-    #                    action='request_reset')
-    #     return urlmap
-    # def after_map(self, urlmap):
-    #     controller = 'ckanext.security.controllers:MFAUserController'
-    #     # Mapping urls for the MFA/TOTP feature
-    #     urlmap.connect('/configure_mfa/{id:.*}/new',
-    #                    controller=controller,
-    #                    action='new')
-    #     urlmap.connect('mfa_configure', '/configure_mfa/{id:.*}',
-    #                    controller=controller,
-    #                    action='configure_mfa')
-    #     urlmap.connect('/api/mfa_login',
-    #                    controller=controller,
-    #                    action='login')
-    #     return urlmap
+
+def login():
+    headers = {'Content-Type': 'application/json'}
+    (status, res_data) = utils.login()
+    make_response((res_data, status, headers))
+
+
+@login_required
+def configure_mfa(id=None):
+    return utils.configure_mfa(id)
+
+
+@login_required
+def new(id=None):
+    return utils.new(id)
+
+
+mfa_user.add_url_rule('/api/mfa_login', view_func=login)
+mfa_user.add_url_rule('/configure_mfa/<id>', view_func=configure_mfa)
+mfa_user.add_url_rule('/configure_mfa/<id>/new', view_func=new)
+
+
+def get_blueprints():
+    return [mfa_user]
+
+
+# Override user edit form template
+user.edit_user_form = u'security/edit_user_form.html'
